backend/requirements.txt
fastapi
uvicorn
motor
pymongo
firecrawl-py
python-dotenv
pydantic[email]
google-generativeai
python-multipart
aiofiles
